requirements.txt
python-dotenv 
openai 
flask
requests
azure.identity